// TechSales-React/Serveur/mongoDB/models/utilisateurv2.ts
import { ObjectId } from "mongodb";
import { ItemPanier } from "../models//itemPanier.js";
import { Adresse } from "../models/adresse.js";

export interface Utilisateurv2 {
  _id?: ObjectId;
  nom: string;
  prenom: string;
  courriel: string;
  motDePasse: string;
  role: string;
  adresse?: Adresse;
  panier: ItemPanier[];
}